+++ src/components/CommitteeTools/CommitteTools.tsx
@@ -1,8 +1,6 @@
-import { useContext, useEffect, useRef, useState } from "react";
+import { useContext } from "react";
 import Decrypt from "./components/Decrypt/Decrypt";
 import classNames from "classnames";
-import Modal from "../Shared/Modal";
-import { useTranslation } from "react-i18next";
 import "./index.scss";
 import Welcome from "./components/Welcome/Welcome";
 import { VaultProvider, VaultContext } from "./store";
@@ -23,16 +21,6 @@ function Content() {
     'committee-tools-wrapper': true,
     'content': true
   });
-
-  // useEffect(() => {
-  //   // vault must be created
-  //   if (!vault?.vault) {
-  //     setShowCreateVault(true)
-  //   } else {
-  //     setShowUnlockVault(true)
-  //   }
-  // }, [vault])
-
 
   return (
     <div className={committeeToolsWrapper}>
