--- src/components/CommitteeTools/components/NewKey/index.scss
@@ -1,1 +1,0 @@
-@import "../../../../styles/constants.scss";
