+++ src/components/CommitteeTools/components/SelectKeyModal/KeyDetails.tsx
@@ -16,9 +16,9 @@ function KeyDetails({ storedKey }: {
     })
     if (privateKey) {
         return (<div>
-            <p>Private Key<CopyToClipboard value={privateKey?.armor()} /></p>
-            <p>Public Key<CopyToClipboard value={privateKey?.toPublic().armor()} /></p>
-            {storedKey.passphrase && <p>Passphrase<CopyToClipboard value={storedKey.passphrase} /></p>}
+            <div>Private Key<CopyToClipboard value={privateKey?.armor()} /></div>
+            <div>Public Key<CopyToClipboard value={privateKey?.toPublic().armor()} /></div>
+            {storedKey.passphrase && <div>Passphrase<CopyToClipboard value={storedKey.passphrase} /></div>}
         </div>)
     } else {
         return (<></>)
