+++ src/components/CommitteeTools/components/SelectKeyModal/SelectKeyModal.tsx
@@ -2,30 +2,27 @@ import { IStoredKey } from "../../../../types/types"
 import { useContext, useState } from "react"
 import Modal from "../../../Shared/Modal"
 import { VaultContext } from "../../store"
-import NewKey from "../NewKey/NewKey"
-import ImportKey from "../NewKey/ImportKey"
 import KeyDetails from "./KeyDetails"
+import GenerateKey from "./GenerateKey"
+import ImportKey from "./ImportKey"
 
 
-const SelectKeyModal = ({ onSelectKey, setShowModal }
+const SelectKeyModal = ({ setShowModal }
     : { onSelectKey: () => any, setShowModal: (show: boolean) => any }) => {
     const [showImport, setShowImport] = useState(false)
     const [showCreate, setShowCreate] = useState(false)
     const vaultContext = useContext(VaultContext)
-    const selectedAlias = useState(null)
     const [showKey, setShowKey] = useState<IStoredKey>()
-    console.log('vault', vaultContext)
     const vault = vaultContext.vault!
 
-    const onAdded = (added: IStoredKey) => {
-
+    const onKeyAdded = (key: IStoredKey) => {
+        setShowModal(false)
     }
 
-
-    return <Modal title="keys" height="fit-content" setShowModal={setShowModal}>
+    return <Modal title="keys" height="fit-content" width="fit-content" setShowModal={setShowModal}>
         {(showCreate || showImport || showKey) ? <>
-            {showCreate && <NewKey onAdded={onAdded} />}
-            {showImport && <ImportKey onAdded={onAdded} />}
+            {showCreate && <GenerateKey onAdded={onKeyAdded} />}
+            {showImport && <ImportKey onAdded={onKeyAdded} />}
             {showKey && <KeyDetails storedKey={showKey} />}
         </> :
             <>
@@ -42,7 +39,7 @@ const SelectKeyModal = ({ onSelectKey, setShowModal }
                                     }}>show</a>
                                     <a href="#" onClick={() => {
                                         vaultContext?.setSelectedAlias!(key.alias)
-                                        onSelectKey()
+                                        setShowModal(false)
                                     }}>select</a>
                                 </div>
                             </li>)}
