+++ src/components/CommitteeTools/index.scss
@@ -29,7 +29,6 @@ $committee-blue: #C0F2F2;
     background-color: $field-blue;
     color: $white;
     border: 1px solid $field-blue;
-    margin-bottom: 10px;
     padding: 15px;
   }
